Allow XML Signatures in one more location on responses
The XML Signature standard is designed to be re-usable standard by
other standards, and it's up to those other standards how they bind it
into their own messages.  In SAML, the binding allows signatures to
appear in a number of different places.  XML Signatures can't assume
anything about the document structure for the standard it is used
with, so instead it uses references to specify which part of the
document the signature applies to.  A given directory may be
consistent about which node it signs, and the previous implementation
was coded to assume the signature appears on the root element.
However, with Azure the "Assertion" block is the signed part.

Update the implementation to delegate responsibility of which node to
pass to "xmlsec" for signature verification to a response method, and
add a basic implementation which supports ID referencing.