AuthnRequest protocol compliance fix
Some implementations might accept xml elements at a particular level
in any order, but due to the design of XMLSchema, typically only one
ordering is valid.  Azure rejects requests with the signature in the
wrong place.  Fix this and another minor compliance issue
(capitalization of an attribute not matching the spec)