Add (limited) support for encrypted response assertions
It's possible to encrypt SAML response assertions; this was primarily
useful in the era where it ran over non-encrypted endpoints, and is
still useful if there is some information in the assertions which you
don't want the user to be able to access using tools such as Firefox
SAML Tracer.  Being an OASIS standard, there are of course several
ways this can be arranged with relation to signatures: the assertions
can be signed and then encrypted, or the assertions can be encrypted
and the whole message signed.

This implementation allows for Assertion nodes to appear under
EncryptedAssertion, which is the form that 'xmlsec1' expects and
returns the plaintext versions, even though it doesn't conform to the
XML schema.  Callers should take utmost care to keep track of which
sections have been validated and which haven't when regarding
assertions as authorative.